LayoutTests/inspector/debugger/bind-script-to-resource.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/debugger-test.js"></script>

<script>

document.write("<scrip" + "t>function foo() { }</sc" + "ript>");

var test = function()
{
    var scripts = [];
    InspectorTest.startDebuggerTest(step1);

    function step1()
    {
        var model = WebInspector.debuggerPresentationModel;

        var scriptURLs = [];
        for (var scriptId in model._rawSourceCodeForScriptId) {
            var script = WebInspector.debuggerModel.scripts[scriptId];
            scriptURLs.push(WebInspector.displayNameForURL(script.sourceURL));
        }
        scriptURLs.sort();
        InspectorTest.assertEquals(4, scriptURLs.length);
        InspectorTest.addResult("There should be 4 scripts:");
        for (var i = 0; i < scriptURLs.length; ++i)
            InspectorTest.addResult("script" + (i+1) + ": " + scriptURLs[i]);

        var urls = Object.keys(model._rawSourceCodeForURL);
        urls.sort();

        InspectorTest.addResult("\nThere should be 3 resources, each resource should have 1 script:");
        InspectorTest.assertEquals(3, urls.length);
        for (var i = 0; i < urls.length; ++i) {
            var url = urls[i];
            var displayName = WebInspector.displayNameForURL(url);
            InspectorTest.addResult("url" + (i+1) + " : " + displayName);
            var rawSourceCode = model._rawSourceCodeForURL[url];
            InspectorTest.assertEquals(1, rawSourceCode._scripts.length);
            for (var j = 0; j < rawSourceCode._scripts.length; ++j) {
                var script = rawSourceCode._scripts[j];
                InspectorTest.addResult("    script: " + WebInspector.displayNameForURL(script.sourceURL));
            }
        }

        urls = Object.keys(model._rawSourceCodeForDocumentURL);
        urls.sort();
        for (var i = 0; i < urls.length; ++i)
            InspectorTest.addResult("document url:" + WebInspector.displayNameForURL(urls[i]));

        InspectorTest.completeDebuggerTest();
    }
}
</script>
</head>

<body onload="runTest()">

<p>
Tests that only inline scripts are bound to the resource, while the ones generated with document.write are not. Bug 71099</a>
</p>

</body>

</html>
