LayoutTests/inspector/debugger/bind-script-to-resource-expected.txt

Tests that only inline scripts are bound to the resource, while the ones generated with document.write are not. Bug 71099

Debugger was enabled.
There should be 4 scripts:
script1: bind-script-to-resource.html
script2: bind-script-to-resource.html
script3: debugger-test.js
script4: inspector-test.js

There should be 3 resources, each resource should have 1 script:
url1 : debugger-test.js
    script: debugger-test.js
url2 : inspector-test.js
    script: inspector-test.js
url3 : bind-script-to-resource.html
    script: bind-script-to-resource.html
document url:bind-script-to-resource.html
Debugger was disabled.
